♻️ refactor : 댓글 서비스/시리얼라이저/테스트 구조 개선 및 역할 분리
