+++ apps/posts/services/__init__.py
@@ -1,0 +1,1 @@
+from .comment_services import validate_comment_create, validate_comment_update
